requirements.txt
aiohttp
pandas
ta
python-telegram-bot==13.15
numpy
python-dotenv